+++ init/src/mod.ts
@@ -1,7 +1,7 @@
 import { ensureDir } from "@std/fs/ensure-dir";
 
 // This value is changed in the release pipeline
-const LIMETTE_VERSION = "0.0.17";
+const LIMETTE_VERSION = "0.0.18";
 
 const projectName = prompt("Your project name?");
 
@@ -36,9 +36,9 @@ const denoJson = `
   },
   "imports": {
     "@limette/core": "jsr:@limette/core@${LIMETTE_VERSION}",
-    "/lit": "npm:/lit@^3.2.0/",
-    "lit": "npm:lit@^3.2.0",
-    "tailwindcss": "npm:tailwindcss@^3.4.14"
+    "/lit": "npm:/lit@^3.2.1/",
+    "lit": "npm:lit@^3.2.1",
+    "tailwindcss": "npm:tailwindcss@^3.4.17"
   },
   "fmt": {
     "singleQuote": true