+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add support for `Middlewares`
 - Add `params` property to LimetteContext
 
 ### Changed